\name{rot90}
\alias{rot90}

\title{Rotate matrix 90 degrees}
\description{Rotate matrix 90 degrees}
 
\usage{
rot90(x)
}

\arguments{
  \item{x}{A square matrix. }
}

\value{ 90 degrees rotation of $x$.
}

\references{Raimondo, M. and Stewart, M. (2007),
`The WaveD Transform in R', Journal of Statistical Software.}
}
\author{Marc Raimondo}

\seealso{\code{\link{WaveD}}}
\examples{ rot90(1:5)}
\keyword{internal}% 
